+++ config/packages/elpa.py
@@ -78,9 +78,11 @@ class Elpa(package.Package):
           f = []
         (result, output) = self.Link([],[],' '.join(l+f),code,' '.join(f),petsc.language)
         if result:
+          self.libflags = ' '.join(l)
+          self.includeflags = ' '.join(f)
           slepcconf.write('#define SLEPC_HAVE_ELPA 1\n')
-          slepcvars.write('ELPA_LIB = ' + ' '.join(l) + '\n')
-          slepcvars.write('ELPA_INCLUDE = ' + ' '.join(f) + '\n')
+          slepcvars.write('ELPA_LIB = ' + self.libflags + '\n')
+          slepcvars.write('ELPA_INCLUDE = ' + self.includeflags + '\n')
           self.havepackage = True
           self.packageflags = ' '.join(l+f)
           return
@@ -124,9 +126,11 @@ class Elpa(package.Package):
       self.log.Exit('Unable to link with downloaded ELPA')
 
     # Write configuration files
+    self.libflags = l
+    self.includeflags = f
     slepcconf.write('#define SLEPC_HAVE_ELPA 1\n')
-    slepcvars.write('ELPA_LIB = ' + l + '\n')
-    slepcvars.write('ELPA_INCLUDE = ' + f + '\n')
+    slepcvars.write('ELPA_LIB = ' + self.libflags + '\n')
+    slepcvars.write('ELPA_INCLUDE = ' + self.includeflags + '\n')
 
     self.havepackage = True
     self.packageflags = l+' '+f