+++ include/slepc/finclude/slepcsvd.h
@@ -27,6 +27,8 @@
 #define SVDStop             PetscEnum
 #define SVDPRIMMEMethod     PetscEnum
 #define SVDTRLanczosGBidiag PetscEnum
+#define SVDKSVDEigenMethod  PetscEnum
+#define SVDKSVDPolarMethod  PetscEnum
 
 #define SVDCROSS      'cross'
 #define SVDCYCLIC     'cyclic'
@@ -35,6 +37,7 @@
 #define SVDTRLANCZOS  'trlanczos'
 #define SVDRANDOMIZED 'randomized'
 #define SVDSCALAPACK  'scalapack'
+#define SVDKSVD       'ksvd'
 #define SVDELEMENTAL  'elemental'
 #define SVDPRIMME     'primme'