+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - New `DSHSVD` type for the hyperbolic SVD.
 - Support for the hyperbolic SVD (HSVD) in the `SVDTRLANCZOS` solver.
+- Add interface to dense SVD solvers in the external package KSVD.
 
 ## [3.18] - 2022-10-01